Made SpecimenAutoV1 work
